requirements.txt
anyio==4.12.1
APScheduler==3.11.2
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
cryptography==46.0.5
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
playwright==1.58.0
pycparser==3.0
pyee==13.0.1
PyJWT==2.12.1
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2026.1.post1
soupsieve==2.8.3
typing_extensions==4.15.0
tzlocal==5.3.1
